implicit uint256 & uint8 on MultiSigWallet contracts and work in progress on tests

+++ contracts/helpers/MultiSigWalletMock.sol
@@ -5,10 +5,10 @@ import '../MultiSigWallet.sol';
 contract MultiSigWalletMock is MultiSigWallet {
     uint256 public transactionId;
 
-    function MultiSigWalletMock(address[] _owners, uint _required) public MultiSigWallet(_owners, _required) {
+    function MultiSigWalletMock(address[] _owners, uint8 _required) public MultiSigWallet(_owners, _required) {
     }
 
-    function submitTransaction(address _destination, uint _value, bytes _data) public returns (uint _transactionId) {
+    function submitTransaction(address _destination, uint256 _value, bytes _data) public returns (uint256 _transactionId) {
         transactionId = super.submitTransaction(_destination, _value, _data);
 
         return transactionId;
